duck_farm_backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.8.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
